+++ src/Providers/BaseProvider.php
@@ -4,7 +4,7 @@ namespace R94ever\CurrencyExchangeRate\Providers;
 
 use R94ever\CurrencyExchangeRate\Enums\Currency;
 
-abstract class BaseProvider
+abstract class BaseProvider implements ExchangeRateProviderInterface
 {
     public function __construct()
     {
@@ -12,4 +12,6 @@ abstract class BaseProvider
     }
 
     abstract public function convert(float $amount, Currency $from, Currency $to): ?float;
+
+    abstract public function getRates(Currency $source, array $targets): array;
 }