+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `php-currency-exchange-rate` will be documented in this file.
 
+## 1.1.0 - 2025-10-05
+
+### Added
+- Added `getRates()` method to get exchange rates for multiple currencies in a single call
+- New `CurrencyRate` value object to represent exchange rates
+- Extended provider interface with `getRates()` method
+- Added comprehensive tests for new functionality
+
 ## 1.0.1 - 2025-10-05
 
 ### Added